requirements.txt
# Django Core
Django==5.2.4
djangorestframework
djangorestframework-simplejwt

# Database
psycopg2-binary

# Cache & Sessions
redis
django-redis

# Additional development tools
django-ratelimit==4.1.0
django-csp==3.7

# PDF Generation (optional)
# weasyprint==60.2

# Additional testing tools
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0 

# Celery
celery==5.3.4
django-celery-beat
django-celery-results

# API Documentation
drf-yasg==1.21.7

# CORS
django-cors-headers==4.3.1

# Internationalization
django-countries==7.6.1
django-phonenumber-field==7.3.0
phonenumbers==9.0.10

# Utilities
django-filter==23.5
django-extensions==3.2.3

# Development & Testing
django-debug-toolbar==4.3.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Environment Variables
python-decouple==3.8

# Image Processing
Pillow

# HTTP Requests
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Security
cryptography==41.0.7
django-axes==6.1.1
bleach==6.1.0
argon2-cffi==25.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk[django]==1.38.0 